core: refactor de carpetas
