+++ src/app/components/bag/bag.component.html
