+++ src/app/components/bag/bag.component.scss
