+++ src/app/components/footer/footer.component.scss
