+++ src/app/components/pokemon-list/pokemon-list.component.scss
