+++ src/app/components/pokemon-list/pokemon-list.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Pokemon } from '../interfaces/Pokemon';
-import { PokemonPoketService } from '../pokemon-poket.service';
-import { PokemonDataService } from '../data/pokemon-data-service.service';
-import { mergeMap, toArray } from 'rxjs/operators';
+import { Pokemon } from '../../interfaces/Pokemon';
+
+import { PokemonPoketService } from '../../services/pokemon-poket.service';
+import { PokemonDataService } from '../../data/pokemon-data-service.service';
 
 @Component({
   selector: 'pokemon-list',
@@ -18,7 +18,7 @@ export class PokemonListComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    if (this.pokemonList.length > 0) return; 
+    if (this.pokemonList.length > 0) return;
     this.pokemonDataService.fetchAllPokemonDetails().subscribe((data) => {
       this.pokemonList = data; // Asignar los datos recopilados a pokemonList
     });
