+++ src/app/components/pokemon-poket-slot/pokemon-poket-slot.component.html
