+++ src/app/components/poket-slot/poket-slot.component.html
