+++ src/app/components/poket/poket.component.scss
