+++ src/app/components/shop-screen/shop-screen.component.ts
