+++ src/app/pages/pokemon-screen/pokemon-screen.component.ts
@@ -3,8 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-pokemon-screen',
   templateUrl: './pokemon-screen.component.html',
-  styleUrl: './pokemon-screen.component.scss'
+  styleUrl: './pokemon-screen.component.scss',
 })
-export class PokemonScreenComponent {
-
-}
+export class PokemonScreenComponent {}
