+++ src/app/pages/pokemon-storage-screen/pokemon-storage-screen.component.html
